update:validate fields, relocate migration file
